Created a collectible.
These are the Seeds of Evil.  Lloyd and Armin are the Watchers.  They're duty is to watch over the Seeds, and make sure that they don't take root, but their evil has seeped into the bugs of the undergrowth, making them aggressive.  It's Lloyd and Armin's duty to recapture the 9 seeds and the Germ of Evil (a coming collectible, and the last one collected to win).